libs/mngr_mind_chat/pyproject.toml
[project]
name = "imbue-mngr-mind-chat"
version = "0.1.0"
description = "Chat command plugin for mngr - connect to mind chat sessions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imbue-mngr==0.1.5",
    "click-option-group>=0.5.6",
    "urwid>=2.2.0",
]

[project.entry-points.mngr]
mind_chat = "imbue.mngr_mind_chat.plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["imbue"]

[tool.uv.sources]
imbue-mngr = { workspace = true }

[tool.pyright]
venvPath = "../.."
venv = ".venv"
pythonVersion = "3.11"
strict = ["**/*.py"]

[tool.pytest.ini_options]
testpaths = ["imbue/mngr_mind_chat"]
python_files = ["*_test.py", "test_*.py"]
python_classes = []
python_functions = ["test_*"]
addopts = [
  "--tb=short",
  "--strict-markers",
  "--timeout=120",
  "--cov=imbue.mngr_mind_chat",
  "--cov-report=term-missing",
  "--cov-fail-under=80",
  "-m not acceptance and not release",
]
markers = [
    "acceptance: acceptance tests (run in CI)",
    "release: release tests (run in CI)",
]

[tool.coverage.run]
source = ["imbue/mngr_mind_chat"]

[tool.coverage.report]
fail_under = 80
omit = [
    "*_test.py",
    "test_*.py",
    "conftest.py",
]
exclude_lines = [
    "pragma: no cover",
]

[[tool.importlinter.contracts]]
name = "mngr_mind_chat layers contract"
type = "layers"
layers = [
    "imbue.mngr_mind_chat.plugin",
    "imbue.mngr_mind_chat.cli",
    "imbue.mngr_mind_chat.api",
]
